pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "disfluentes"
version = "0.1.0"
description = "A Python tool for generating natural disfluencies in Spanish text"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["spanish", "nlp", "disfluency", "text-generation"]
authors = [
    { name = "LIAA Research Group" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Spanish",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "spacy>=3.7.2",
    "spacy-spanish-lemmatizer>=0.5",
    "fire>=0.5.0",
    "gin-config>=0.5.0",
    "ipython>=8.12.0",
    "pytest>=7.4.0",
    "mlconjug3>=3.10.3",
    "spacy_syllables",
]

[project.scripts]
disfluentes = "disfluenciate:main"


[project.urls]
Homepage = "https://github.com/habla-liaa/disfluentES"
Repository = "https://github.com/habla-liaa/disfluentES"
Issues = "https://github.com/habla-liaa/disfluentES/issues"

[tool.hatch.build.targets.wheel]
packages = ["disfluentes"]


[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
] 